refactor: Home 경로 폴더 구조 변경
