+++ src/pages/Home/components/TimeSelector/TimeCarousel/style.ts
