+++ src/pages/Home/components/WeatherCard/style.ts
