+++ src/routes.tsx
@@ -3,7 +3,7 @@ import Home from './pages/Home';
 import NotFound from './pages/NotFound';
 import RootLayout from './components/layout/RootLayout';
 import Search from './pages/Search';
-import WeatherCard from './pages/Home/components/weather/WeatherCard';
+
 export const router = createBrowserRouter([
   {
     element: <RootLayout />,
@@ -14,7 +14,6 @@ export const router = createBrowserRouter([
         path: '/search',
         element: <Search />,
       },
-      { path : '/weathercard', element: <WeatherCard />},
     ],
   },
 ]) as ReturnType<typeof createBrowserRouter>;
